Added initial Go implementation for the API service
I have decided to rewrite the project in Go as a mean to improve my
skills in the language, this was originally written as a web application
in PHP then rewritten into Node.JS and recently after getting more
interested in the Go programming language I have decided to use it more
in my personal projects. Hopefully this will be both an API service and
a command line tool to send content to the repository.