Added support to render static files without dir listing
CSS, JavaScript and font files stored in the assets directory will be
available in the endpoint "/assets/*". However, the http.FileServer does
automatic directory listing which is not good from a security point of
view, so this commit also adds a mechanism to prevent this when the URL
ends in forward slash instead of a file name.