Added HTML interface with stylesheet and font files
This commit adds the necessary HTML code to render the interface with
the ACE editor and some custom elements in the bottom of the pane that
will act as a status bar, and the top of the pane that will act as the
toolbar. This also includes the required CSS and font files.