Added new screenshot and deleted unnecessary files
This adds a new image showcasing the new interface after the initial
rewrite of the code, additionally to the deletion of unnecessary HTML
files that were being used by the old Node.JS-based web application.
Notice that the functionality is still broken but it will be added in
subsequent commits.