Add API endpoint to write content into the service
The API service will receive a POST request with parameters "mode" and
"content" with the name of one of the supported syntax highlighters and
the data respectively. A 400 Bad Request HTTP response will be returned
if either the mode is not supported or one of the required parameters is
missing. Notice that the generation of the random unique filename takes
in consideration the first character in the string to link the file into
a specific sub-directory, this helps to keep the size of the storage
folder in manageable limits.