Add API endpoint to list supported syntax highlighters
The user can change the syntax highlight of the code editor by clicking
one of the supported programming languages in the syntax section. This
will retrieve an updated list of the ACE JavaScript library modes. Note
that if you are going to add a new entry in the list you have to restart
the server.