Add custom ServeHTTP to leverage routing to interface
A custom router will attach the defined HTTP handlers into a list of
nodes, if during the request the URL path is found in this list the
custom server will execute then associated handler, otherwise will
return a 404 Not Found HTTP status. If the user sends a request with
an unsupported HTTP method the server will return a 405.