Add support for static files via custom HTTP server
The developer can define a GET and/or POST request to handle requests
directed to a static file located one or more directories. The API will
check the URL against the defined paths and check the existence of an
attribute named Node.MatchEverything in which case will skip the dynamic
named parameter processing and immediately call the http.FileServer.